User menus

+++ src/app/store/user/actions.js
@@ -9,21 +9,28 @@ import { jwtSelector } from './selectors';
 export const SIGN_IN = 'SIGN_IN';
 export const SIGN_IN_SUCCESFULL = 'SIGN_IN_SUCCESFULL';
 export const SET_CURRENT_USER = 'SET_CURRENT_USER';
-
 export const SIGN_OUT = 'SIGN_OUT';
+export const SET_USER_SANDBOXES = 'SET_USER_SANDBOXES';
 
 export const GET_CURRENT_USER_API = createAPIActions('CURRENT_USER', 'FETCH');
+export const LOAD_USER_SANDBOXES = createAPIActions(
+  'CURRENT_USER',
+  'FETCH_SANDBOXES'
+);
 
 const deleteCookie = (name: string) => {
   document.cookie = `${name}=; Path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT;`;
 };
 
 const signOut = () => async (dispatch: Function) => {
-  deleteCookie('jwt');
+  const yes = confirm('Are you sure you want to sign out?');
+  if (yes) {
+    deleteCookie('jwt');
 
-  dispatch({
-    type: SIGN_OUT,
-  });
+    dispatch({
+      type: SIGN_OUT,
+    });
+  }
 };
 
 const getCurrentUser = () => async (dispatch: Function, getState: Function) => {
@@ -65,8 +72,18 @@ const signIn = () => async (dispatch: Function) => {
   });
 };
 
+const loadUserSandboxes = () => async (dispatch: Function) => {
+  const { data } = await dispatch(doRequest(LOAD_USER_SANDBOXES, `sandboxes`));
+
+  dispatch({
+    type: SET_USER_SANDBOXES,
+    data,
+  });
+};
+
 export default {
   signOut,
   signIn,
   getCurrentUser,
+  loadUserSandboxes,
 };
